HW_1/requirements.txt
asttokens==3.0.1
comm==0.2.3
contourpy==1.3.3
cuda-bindings==13.0.3
cuda-pathfinder==1.2.2
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
ipykernel==7.2.0
ipython==9.10.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.8.0
jupyter-core==5.9.1
kiwisolver==1.4.9
markupsafe==2.1.5
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.3.5
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.15.1.9
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
parso==0.8.6
pexpect==4.9.0
pillow==12.0.0
pip @ file:///home/conda/feedstock_root/build_artifacts/pip_1770270195740/work
platformdirs==4.9.2
prompt-toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyzmq==27.1.0
setuptools==82.0.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
torch==2.10.0+cu130
torchvision==0.25.0+cu130
tornado==6.5.4
traitlets==5.14.3
triton==3.6.0
typing-extensions==4.15.0
uv==0.10.4
wcwidth==0.6.0
wheel==0.46.3
